Remove clamp call from dwell handling.
This removes the clamp() operation from the dwell instrumentation. This
was effectively enforcing a minimum dwell time of 1 second, which
doesn't make sense in most situation, especially with a machine
optimized for drill/tap operations.

The cycle call was effectively creating a min/max handling, but by
enforcing a 1 second minimum, it could add up on cycles involving a lot
of dwells that had their settings to something less than 1 second.