Remove vertical stock to leave from smoothing
This removes the checking of vertical stock to leave from the smoothing
threshold handling. The way it was instrumented required *both* axial
and radial stock to be above the reshold in order for it be considered
roughing.

Most of the time, I do not make heavy use of vertical stock to leave in
roughing operations when doing outside clearing. When fixturing, I hog
off everything. When doing vise work, I'm often going below the model,
either with negative axial or a negative bottom height.

I may later change it to add vertical back and use it as a check for
mild roughing vs full roughing, but for now will run it without to see.