Fix probing of Z with Blum probe
This fixes an issue with probing Z when using a Blum probe. By default,
the Blum macros expect relative distances to be passed, howevever for it
to accept an absolute position, the "A1" parameter needs to be passed.

Note that I didn't want to change it to pass a relative distance,
because that could interfere with doing tolerance inspections on
non-zero locations, such as probe something thats 1", ensure its within
.004" tolerance. It would zero to the top, and alarm saying over
tolerance by 1".

Another approach could maybe be to pass relative Z, but add K with
absolute position.