Added support for inline tool break detection
This adds the proper tool break detection, keyed off the "break control"
checkbox in the tool library entry.

It will call out to the standard 5000 macro on the Speedio to do the
break detection. On my machine, the macro itself does the Z retraction.

If there are multiple toolpaths using the same tool that has tool break
detection enabled, it will only perform the check once at the end of the
full set.

Note that currently, the way this is `M05` will be inserted twice. After
the macro sub, it will have `M05` again because of the standard call in
`onSelection` for a tool change.